shift bulk upload CTA to inline callout
